+++ tests/test_models.py
@@ -133,6 +133,18 @@ class BarTimeTestCase(TestCase):
         k = BarTime.objects.filter(datetime='1398-04-31 12:12:12')
         self.assertEqual(k[0].datetime.day, 31)
 
+    def test_date_lookup_filter(self):
+        jdatetime_ = jdatetime.datetime(1378, 10, 10)
+        bartime = BarTime.objects.create(name="Test", datetime=jdatetime_)
+        self.assertEqual(BarTime.objects.get(datetime__date=jdatetime_), bartime)
+        self.assertEqual(BarTime.objects.get(datetime__date=jdatetime_.date()), bartime)
+        self.assertEqual(BarTime.objects.get(datetime__date="1378-10-10"), bartime)
+        with self.assertRaisesMessage(
+            TypeError,
+            "`__date` filter Expected jdatetime.datetime, jdatetime.date or str, got `int`.",
+        ):
+            BarTime.objects.get(datetime__date=1)
+
     def test_filter_by_exact_datetime(self):
         bar_times = BarTime.objects.filter(datetime=self.datetime)
         self.assertEqual(len(bar_times), 1)