+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Unreleased
+- Add support for `__date` lookup filter
+
 ### Changed
 - Remove deprecated `datetime_safe` and use `sanitize_strftime_format`